INSERT/UPDATE/SELECT/DELETE working without prepending schema
Right now there are 3 tests failures due to the fact that
the schema.sql fixture is overriding the json type with a custom table
named "json".(Specifically the error comes from this part of the
produced query: `json_populate_record(NULL:: "json"`)

This could be fixed by doing a `SET search_path TO test, pg_catalog`
to show a preference for the test schema.

However looking at the code history there's not a strong reason
to keep the "json" fixture, it would be better to just rename it
to perhaps a "json_data" TABLE.

This json TABLE also cause some issues during testing because the json type
has to be used as "pg_catalog.json" giving some confusing error
messages.